Add clear-effects-on-load config option (#17)
# Description

Added a boolean option to toggle the cleanse of all effects upon loading
the resourcepack. This change keeps the cleaning of the Blindness
effect. Also edited the disableEffects method, to properly accomodate
the config changes.

## Type of change

Please delete options that are not relevant.

- [x] New feature (non-breaking change which adds functionality)
- [x] This change requires a documentation update

# How Has This Been Tested?

Changes were tested locally and on a hosted server, and proven to work
as intended

# Checklist:

- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my code
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation, if any
- [x] My changes generate no new warnings
- [x] I have performed tests that prove my fix is effective or that my
feature works, if necessary
- [ ] New and existing unit tests pass locally with my changes

Co-authored-by: Ian Tapply <iantapply22@gmail.com>